requirements.txt
python-dotenv
undetected-chromedriver
selenium